+++ lib/project_percent_done/calculation_result.rb
@@ -18,12 +18,19 @@ module ProjectPercentDone
                 :unestimated_eligible_issue_count,
                 :included_issue_count,
                 :excluded_parent_issue_count,
+                :excluded_non_progress_issue_count,
+                :excluded_non_progress_estimated_hours,
+                :excluded_non_progress_applied_weight,
+                :non_progress_status_ids,
+                :non_progress_tracker_ids,
+                :non_progress_status_names,
+                :non_progress_tracker_names,
                 :ignored_unestimated_issue_count,
                 :known_estimated_hours,
                 :imputed_weight,
                 :estimate_coverage_percent
 
-    def initialize(percent_done:, raw_percent_done:, issue_count:, not_included_issue_count:, estimated_issue_count:, unestimated_issue_count:, total_weight:, included_issue_ids:, not_included_issue_ids:, included_rows: [], not_included_rows: [], warnings: [], all_project_issue_count: 0, eligible_issue_count: 0, estimated_eligible_issue_count: 0, unestimated_eligible_issue_count: 0, included_issue_count: 0, excluded_parent_issue_count: 0, ignored_unestimated_issue_count: 0, known_estimated_hours: 0.0, imputed_weight: 0.0, estimate_coverage_percent: nil)
+    def initialize(percent_done:, raw_percent_done:, issue_count:, not_included_issue_count:, estimated_issue_count:, unestimated_issue_count:, total_weight:, included_issue_ids:, not_included_issue_ids:, included_rows: [], not_included_rows: [], warnings: [], all_project_issue_count: 0, eligible_issue_count: 0, estimated_eligible_issue_count: 0, unestimated_eligible_issue_count: 0, included_issue_count: 0, excluded_parent_issue_count: 0, excluded_non_progress_issue_count: 0, excluded_non_progress_estimated_hours: 0.0, excluded_non_progress_applied_weight: 0.0, non_progress_status_ids: [], non_progress_tracker_ids: [], non_progress_status_names: [], non_progress_tracker_names: [], ignored_unestimated_issue_count: 0, known_estimated_hours: 0.0, imputed_weight: 0.0, estimate_coverage_percent: nil)
       @percent_done = percent_done
       @raw_percent_done = raw_percent_done
       @issue_count = issue_count
@@ -42,6 +49,13 @@ module ProjectPercentDone
       @unestimated_eligible_issue_count = unestimated_eligible_issue_count
       @included_issue_count = included_issue_count
       @excluded_parent_issue_count = excluded_parent_issue_count
+      @excluded_non_progress_issue_count = excluded_non_progress_issue_count
+      @excluded_non_progress_estimated_hours = excluded_non_progress_estimated_hours
+      @excluded_non_progress_applied_weight = excluded_non_progress_applied_weight
+      @non_progress_status_ids = non_progress_status_ids
+      @non_progress_tracker_ids = non_progress_tracker_ids
+      @non_progress_status_names = non_progress_status_names
+      @non_progress_tracker_names = non_progress_tracker_names
       @ignored_unestimated_issue_count = ignored_unestimated_issue_count
       @known_estimated_hours = known_estimated_hours
       @imputed_weight = imputed_weight
@@ -68,6 +82,13 @@ module ProjectPercentDone
         :unestimated_eligible_issue_count => unestimated_eligible_issue_count,
         :included_issue_count => included_issue_count,
         :excluded_parent_issue_count => excluded_parent_issue_count,
+        :excluded_non_progress_issue_count => excluded_non_progress_issue_count,
+        :excluded_non_progress_estimated_hours => excluded_non_progress_estimated_hours,
+        :excluded_non_progress_applied_weight => excluded_non_progress_applied_weight,
+        :non_progress_status_ids => non_progress_status_ids,
+        :non_progress_tracker_ids => non_progress_tracker_ids,
+        :non_progress_status_names => non_progress_status_names,
+        :non_progress_tracker_names => non_progress_tracker_names,
         :ignored_unestimated_issue_count => ignored_unestimated_issue_count,
         :known_estimated_hours => known_estimated_hours,
         :imputed_weight => imputed_weight,