+++ test/functional/project_percent_done_controller_test.rb
@@ -28,10 +28,15 @@ class ProjectPercentDoneControllerTest < ActionController::TestCase
       assert_equal test_project.identifier, payload['project_percent_done']['project_identifier']
       assert_equal(
         %w[
-          calculation_mode closed_issue_mode estimated_issue_count issue_count
-          not_included_issue_count percent_done project_id project_identifier
-          raw_percent_done total_weight unestimated_issue_count
-          unestimated_issue_mode warnings
+          calculation_mode closed_issue_mode estimated_issue_count
+          excluded_non_progress_applied_weight
+          excluded_non_progress_estimated_hours excluded_non_progress_issue_count
+          issue_count non_progress_status_ids non_progress_status_names
+          non_progress_status_scope non_progress_tracker_ids
+          non_progress_tracker_names non_progress_tracker_scope
+          not_included_issue_count percent_done
+          project_id project_identifier raw_percent_done total_weight
+          unestimated_issue_count unestimated_issue_mode warnings
         ],
         payload['project_percent_done'].keys.sort
       )
@@ -90,6 +95,88 @@ class ProjectPercentDoneControllerTest < ActionController::TestCase
       )
       assert_includes @response.body, I18n.t(:label_project_percent_done_history_chart_mode_control)
       refute_includes @response.body, I18n.t(:label_project_percent_done_history_chart_mode)
+    end
+  end
+
+  def test_calculation_details_render_issue_table_filters
+    parent = create_test_issue(:subject => 'Parent task', :done_ratio => 100, :estimated_hours => 20)
+    create_test_issue(:subject => 'Included child task', :parent_issue_id => parent.id, :done_ratio => 50, :estimated_hours => 10)
+
+    with_project_percent_done_settings('display_project_tab' => '1') do
+      get :show, :params => { :project_id => test_project.identifier }
+
+      assert_response :success
+      assert_select 'div.box.ppd-issue-list-section.ppd-collapsible-section', :count => 2
+      assert_select 'div.ppd-issue-list-section[data-ppd-section="included"]' do
+        assert_select 'a.icon.icon-save[href*=".csv"][href*="table=included"]', :text => 'CSV'
+        assert_select 'button.ppd-section-toggle[data-ppd-toggle-section="included"][aria-expanded="false"]',
+                      :text => /#{Regexp.escape(I18n.t(:label_project_percent_done_included_issues))}/
+        assert_select 'div.ppd-collapsible-body[data-ppd-section-body="included"][hidden]'
+      end
+      assert_select 'div.ppd-issue-list-section[data-ppd-section="not-included"]' do
+        assert_select 'a.icon.icon-save[href*=".csv"][href*="table=not_included"]', :text => 'CSV'
+        assert_select 'button.ppd-section-toggle[data-ppd-toggle-section="not-included"][aria-expanded="false"]',
+                      :text => /#{Regexp.escape(I18n.t(:label_project_percent_done_not_included_issues))}/
+        assert_select 'div.ppd-collapsible-body[data-ppd-section-body="not-included"][hidden]'
+      end
+      assert_select 'div.ppd-table-filter[data-ppd-filter-controls="ppd-included-issues"]' do
+        assert_select 'label span', :text => I18n.t(:label_project_percent_done_filter)
+        assert_select 'input[type="search"][data-ppd-filter-search]'
+        assert_select 'select[data-ppd-filter-status]'
+        assert_select 'select[data-ppd-filter-secondary]'
+        assert_select 'button[data-ppd-quick-filter="closed_as_100"]'
+        assert_select 'span[data-ppd-filter-count]'
+      end
+      assert_select 'table#ppd-included-issues[data-ppd-filter-table]' do
+        assert_select 'th.ppd-issue-subject', :text => I18n.t(:field_subject)
+        assert_select 'td.ppd-issue-subject'
+        assert_select 'tr[data-ppd-filter-row][data-ppd-search][data-ppd-status][data-ppd-secondary][data-ppd-quick]'
+        assert_select 'tr[data-ppd-filter-empty][hidden]'
+      end
+      assert_select 'div.ppd-table-filter[data-ppd-filter-controls="ppd-not-included-issues"]' do
+        assert_select 'label span', :text => I18n.t(:label_project_percent_done_filter)
+        assert_select 'button[data-ppd-quick-filter="parent_issue_excluded"]'
+        assert_select 'button[data-ppd-quick-filter="non_progress_status"]'
+        assert_select 'span[data-ppd-filter-count]'
+      end
+      assert_select 'table#ppd-not-included-issues[data-ppd-filter-table]' do
+        assert_select 'th.ppd-issue-subject', :text => I18n.t(:field_subject)
+        assert_select 'td.ppd-issue-subject'
+        assert_select 'tr[data-ppd-filter-row][data-ppd-quick="parent_issue_excluded"]'
+      end
+      assert_includes @response.body, 'refreshFilter'
+      assert_includes @response.body, 'setSectionExpanded'
+      assert_includes @response.body, I18n.t(:text_project_percent_done_filter_placeholder)
+    end
+  end
+
+  def test_calculation_details_csv_exports_visible_included_issue_rows
+    create_test_issue(:subject => 'Included CSV task', :done_ratio => 50, :estimated_hours => 10)
+
+    with_project_percent_done_settings('display_project_tab' => '1') do
+      get :show, :params => { :project_id => test_project.identifier, :format => 'csv', :table => 'included' }
+
+      assert_response :success
+      assert_equal 'text/csv', @response.media_type
+      assert_includes @response.headers['Content-Disposition'], 'project-percent-done'
+      assert_includes @response.body, I18n.t(:field_subject)
+      assert_includes @response.body, 'Included CSV task'
+      assert_includes @response.body, '50.00%'
+    end
+  end
+
+  def test_calculation_details_csv_exports_visible_not_included_issue_rows
+    parent = create_test_issue(:subject => 'Parent CSV task', :done_ratio => 100, :estimated_hours => 20)
+    create_test_issue(:subject => 'Child CSV task', :parent_issue_id => parent.id, :done_ratio => 50, :estimated_hours => 10)
+
+    with_project_percent_done_settings('display_project_tab' => '1') do
+      get :show, :params => { :project_id => test_project.identifier, :format => 'csv', :table => 'not_included' }
+
+      assert_response :success
+      assert_equal 'text/csv', @response.media_type
+      assert_includes @response.body, I18n.t(:label_project_percent_done_reason)
+      assert_includes @response.body, 'Parent CSV task'
+      assert_includes @response.body, I18n.t(:label_project_percent_done_reason_parent_issue_excluded)
     end
   end